Streamlined is_it_time() helper function a bit.

+++ Blink_millis6.ino
@@ -34,8 +34,8 @@ uint32_t blink_time;                  // Global variable to store the time that 
 ** time, t0, plus a given interval, dt.  The condition accounts for timer overflow / wraparound.
 */
 bool it_is_time(uint32_t t, uint32_t t0, uint16_t dt) {
-  return ((t >= t0) && (t - t0 >= dt)) ||                   // The first disjunct handles the normal case
-            ((t < t0) && (t + (uint32_t)(~t0) + 1 >= dt));  //   while the second handles the overflow case
+  return ((t >= t0) && (t - t0 >= dt)) ||         // The first disjunct handles the normal case
+            ((t < t0) && (t + (~t0) + 1 >= dt));  //   while the second handles the overflow case
 }
 
 void setup() {
